use utc dayjs plugin to get current time in utc
if not currently in utc, the dayjs function was returning the wrong timestamp
for beginning the day. it still had my timezone as part of it. this update
uses the utc dayjs plugin to ensure we are using a utc-based comparison since
event.created_at in nostr notes are supposed to be in utc time.